Merged from 1_3_bugfix Changed the rounding implementation to NOT use the java6 specific api call DecimalFormat.setRoundingMode(). Instead we do the rounding using BigDecimal, BEFORE letting DecimalFormat do the string formatting.
This required me to restrict a bit in the generification, as we need to know T as being something that extends java.lang.Number. Shouldn't be that bad, as this is just stating the obvious (aka: it needed to be a Number in the first place, because using DecimalFormat has some switching logic deep down in its stomach, that throws up if the given Object to format isn't some kind of a Number). [from revision 162]